BREAKING CHANGE: Add 'attributes' argument to plugins calls.
The new 'attributes' argument contains a dictionary of attributes provided in the task's configuration, allowing for key-value pairs to be defined the task and used by custom plugins.

Previous examples of plugins used a specific number of keyword arguments, meaning that the addition of a new argument would break existing plugins.
New recommendations are to use **kwargs in all method definitions to allow for future kwargs to be added to the calls.

Example notification plugin has been updated with the new arguments and slack example plugin has been updated to use an attribute for notification slack channel, using an environment variable for slack token, which allows the slack app API to be used, instead of the webhook

Issue #58